cleaning out redux store
