+++ src/types/BookStoreTypes/BookStoreTypes.tsx
@@ -1,0 +1,8 @@
+import { Book } from "../../types/Book/Book";
+
+export interface BookState {
+  books: Book[];
+  modalChecker: boolean;
+  modalCheckerModify: boolean;
+  status: string;
+}
